requirements.txt
Flask
Clickgit remote add origin